and fix long->int too

+++ matsim/Network.py
@@ -125,7 +125,7 @@ def read_network(filename, skip_attributes=False):
                 # TODO: pandas will make the value column "object" since we're mixing types
                 if 'class' in elem.attrib:
                     if elem.attrib['class'] == 'java.lang.Long':
-                        atts['value'] = float(elem.text)
+                        atts['value'] = int(elem.text)
                     if elem.attrib['class'] == 'java.lang.Double':
                         atts['value'] = float(elem.text)
                     if elem.attrib['class'] == 'java.lang.Integer':
